// src/types/prisma/schemas/objects/SeriousnessWhereInput.schema.ts
import * as z from 'zod';
import type { Prisma } from '@prisma/client';
import { IntFilterObjectSchema as IntFilterObjectSchema } from './IntFilter.schema';
import { StringFilterObjectSchema as StringFilterObjectSchema } from './StringFilter.schema';
import { DateTimeFilterObjectSchema as DateTimeFilterObjectSchema } from './DateTimeFilter.schema';
import { OccurenceListRelationFilterObjectSchema as OccurenceListRelationFilterObjectSchema } from './OccurenceListRelationFilter.schema'

const seriousnesswhereinputSchema = z.object({
  AND: z.union([z.lazy(() => SeriousnessWhereInputObjectSchema), z.lazy(() => SeriousnessWhereInputObjectSchema).array()]).optional(),
  OR: z.lazy(() => SeriousnessWhereInputObjectSchema).array().optional(),
  NOT: z.union([z.lazy(() => SeriousnessWhereInputObjectSchema), z.lazy(() => SeriousnessWhereInputObjectSchema).array()]).optional(),
  id: z.union([z.lazy(() => IntFilterObjectSchema), z.number().int()]).optional(),
  level: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  createdAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional(),
  updatedAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional(),
  occurences: z.lazy(() => OccurenceListRelationFilterObjectSchema).optional()
}).strict();
export const SeriousnessWhereInputObjectSchema: z.ZodType<Prisma.SeriousnessWhereInput> = seriousnesswhereinputSchema as unknown as z.ZodType<Prisma.SeriousnessWhereInput>;
export const SeriousnessWhereInputObjectZodSchema = seriousnesswhereinputSchema;